+++ packages/xrpl/src/Wallet/utils.ts
@@ -10,6 +10,7 @@ import {
   encode,
   encodeForMultisigning,
   encodeForSigning,
+  XrplDefinitionsBase,
 } from 'ripple-binary-codec'
 import { sign } from 'ripple-keypairs'
 
@@ -61,20 +62,25 @@ export function addressToBigNumber(address: string): BigNumber {
  * Decodes a transaction or transaction blob into a Transaction object.
  *
  * @param txOrBlob - A Transaction object or a hex string representing a transaction blob.
+ * @param definitions - Custom rippled types to use instead of the default. Used for sidechains and amendments.
  * @returns A Transaction object.
  * @throws If the input is not a valid Transaction or transaction blob.
  */
 export function getDecodedTransaction(
   txOrBlob: Transaction | string,
+  definitions?: XrplDefinitionsBase,
 ): Transaction {
   if (typeof txOrBlob === 'object') {
     // We need this to handle X-addresses in multisigning
     // eslint-disable-next-line @typescript-eslint/consistent-type-assertions -- We are casting here to get strong typing
-    return decode(encode(txOrBlob)) as unknown as Transaction
+    return decode(
+      encode(txOrBlob, definitions),
+      definitions,
+    ) as unknown as Transaction
   }
 
   // eslint-disable-next-line @typescript-eslint/consistent-type-assertions -- We are casting here to get strong typing
-  return decode(txOrBlob) as unknown as Transaction
+  return decode(txOrBlob, definitions) as unknown as Transaction
 }
 
 /**
@@ -84,19 +90,25 @@ export function getDecodedTransaction(
  * @param privateKey - A key to sign the transaction with.
  * @param signAs - Multisign only. An account address to include in the Signer field.
  * Can be either a classic address or an XAddress.
+ * @param definitions Custom rippled types to use instead of the default. Used for sidechains and amendments.
  * @returns A signed transaction in the proper format.
  */
+// eslint-disable-next-line max-params -- Needs 4 params
 export function computeSignature(
   tx: Transaction,
   privateKey: string,
   signAs?: string,
+  definitions?: XrplDefinitionsBase,
 ): string {
   if (signAs) {
     const classicAddress = isValidXAddress(signAs)
       ? xAddressToClassicAddress(signAs).classicAddress
       : signAs
 
-    return sign(encodeForMultisigning(tx, classicAddress), privateKey)
+    return sign(
+      encodeForMultisigning(tx, classicAddress, definitions),
+      privateKey,
+    )
   }
-  return sign(encodeForSigning(tx), privateKey)
+  return sign(encodeForSigning(tx, definitions), privateKey)
 }