+++ packages/xrpl/src/models/transactions/common.ts
@@ -1,7 +1,7 @@
 /* eslint-disable max-lines -- common utility file */
 import { HEX_REGEX } from '@xrplf/isomorphic/utils'
 import { isValidClassicAddress, isValidXAddress } from 'ripple-address-codec'
-import { TRANSACTION_TYPES } from 'ripple-binary-codec'
+import { DEFAULT_DEFINITIONS, XrplDefinitionsBase } from 'ripple-binary-codec'
 
 import { ValidationError } from '../../errors'
 import {
@@ -542,11 +542,13 @@ export interface BaseTransaction extends Record<string, unknown> {
  * any time a transaction will be verified.
  *
  * @param common - An interface w/ common transaction fields.
+ * @param definitions - Custom rippled types to use instead of the default. Used for sidechains and amendments.
  * @throws When the common param is malformed.
  */
 // eslint-disable-next-line max-statements, max-lines-per-function -- lines required for validation
 export function validateBaseTransaction(
   common: unknown,
+  definitions: XrplDefinitionsBase = DEFAULT_DEFINITIONS,
 ): asserts common is BaseTransaction {
   if (!isRecord(common)) {
     throw new ValidationError(
@@ -562,7 +564,7 @@ export function validateBaseTransaction(
     throw new ValidationError('BaseTransaction: TransactionType not string')
   }
 
-  if (!TRANSACTION_TYPES.includes(common.TransactionType)) {
+  if (!definitions.transactionNames.includes(common.TransactionType)) {
     throw new ValidationError(
       `BaseTransaction: Unknown TransactionType ${common.TransactionType}`,
     )