+++ packages/xrpl/test/fixtures/requests/index.ts
@@ -4,6 +4,7 @@ import header from './hashLedger.json'
 import transactions from './hashLedgerTransactions.json'
 import normalSign from './sign.json'
 import signAsSign from './signAs.json'
+import signAsCustomDefinition from './signAsCustomDefinition.json'
 import escrowSign from './signEscrow.json'
 import signPaymentChannelClaim from './signPaymentChannelClaim.json'
 import ticketSign from './signTicket.json'
@@ -13,6 +14,7 @@ const sign = {
   ticket: ticketSign,
   escrow: escrowSign,
   signAs: signAsSign,
+  signAsCustomDefinition,
 }
 
 const getOrderbook = {