+++ packages/xrpl/test/fixtures/rippled/index.ts
@@ -5,6 +5,7 @@ import normalAccountTx from './accountTx.json'
 import fabric from './bookOffers'
 import usd_xrp from './bookOffersUsdXrp.json'
 import xrp_usd from './bookOffersXrpUsd.json'
+import customDefinition from './customDefinition.json'
 import normalLedger from './ledger.json'
 import firstPage from './ledgerDataFirstPage.json'
 import firstPageEmpty from './ledgerDataFirstPageEmpty.json'
@@ -96,6 +97,10 @@ const server_info = {
   withNetworkId: withNetworkIDServerInfo,
 }
 
+const definitions = {
+  customDefinition,
+}
+
 const tx = {
   NFTokenMint,
   NFTokenMint2,
@@ -120,6 +125,7 @@ const rippled = {
   transaction_entry,
   tx,
   unsubscribe,
+  definitions,
 }
 
 export default rippled